+++ core/src/it/units/sdm/project/game/terminal/FreedomGame.java
@@ -4,6 +4,7 @@ import it.units.sdm.project.board.Position;
 import it.units.sdm.project.board.Stone;
 import it.units.sdm.project.enums.GameStatus;
 import it.units.sdm.project.game.FreedomPointsCounter;
+import it.units.sdm.project.game.Move;
 import it.units.sdm.project.game.Player;
 import it.units.sdm.project.interfaces.Board;
 import org.jetbrains.annotations.NotNull;
@@ -11,7 +12,6 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.LinkedList;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -193,40 +193,5 @@ public class FreedomGame {
             return blackPlayer;
         }
         return null;
-    }
-
-
-    private static class Move {
-
-        @NotNull
-        private final Player player;
-        @NotNull
-        private final Position position;
-
-        public Move(@NotNull Player player, @NotNull Position position) {
-            this.player = player;
-            this.position = position;
-        }
-
-        public @NotNull Player getPlayer() {
-            return player;
-        }
-
-        public @NotNull Position getPosition() {
-            return position;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            Move move = (Move) o;
-            return player.equals(move.player) && position.equals(move.position);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(player, position);
-        }
     }
 }
